backend/requirements.txt
asgiref==3.2.10
boto3==1.14.9
botocore==1.17.9
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
coreapi==2.3.3
coreapi-cli==1.0.9
coreschema==0.0.4
dj-database-url==0.5.0
Django==3.0.7
django-cors-headers==3.4.0
django-cors-middleware==1.5.0
django-heroku==0.3.1
django-storages==1.9.1
djangorestframework==3.11.0
docutils==0.15.2
gunicorn==20.0.4
hal-codec==1.0.2
idna==2.9
itypes==1.2.0
Jinja2==2.11.2
jmespath==0.10.0
jsonhyperschema-codec==1.0.3
MarkupSafe==1.1.1
openapi-codec==1.3.2
Pillow==7.1.2
psycopg2==2.8.5
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.9
whitenoise==5.1.0
